Пофиксил версию на гите

+++ src/main/java/com/sas/alex/repository/RoleRepository.java
@@ -2,13 +2,14 @@ package com.sas.alex.repository;
 
 import com.sas.alex.model.ERole;
 import com.sas.alex.model.Role;
+import io.swagger.models.auth.In;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
 
 @Repository
-public interface RoleRepository extends JpaRepository<Role, Long> {
+public interface RoleRepository extends JpaRepository<Role, Integer> {
     Optional<Role> findByName(ERole name);
 
 }
